requirements.txt
PyQt5==5.15.10
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
google-generativeai==0.4.0
pandas==2.1.3
plotly==5.17.0
python-dateutil==2.8.2